dts_to_esc: move the member-slot and decl-name helpers out of rerun.go
Refs #1342.

`memberSlot`, `classElemSlot`, `objElemSlot`, `slotFor`, `escDeclName`,
and `escDeclKind` address a declaration by name and a member by slot.
The additive re-run is one caller; the overlay is about to be another,
and #1345 removes `rerun.go` whole. Moving them to `decl_key.go` keeps
that removal a deletion and nothing more.

The comments describe what the helpers do rather than what the re-run
does with them.
